+++ nodeserver/mqttHandler.js
@@ -2,7 +2,7 @@ const { connectToMongo } = require('../nodebackend/build/clients/mongoClient');
 const EventEmitter = require('events');
 const { writeToInflux } = require('./influxDbClient');
 const mqttClient = require('../nodebackend/build/clients/mqttClient').default;
-const { broadcastToSseClients, addSseClient } = require('./sseHandler');
+const { broadcastToSseClients, addSseClient } = require('../nodebackend/build/utils/sseHandler');
 const logger = require('../nodebackend/build/logger').default;
 
 class StateChangeEmitter extends EventEmitter { }