+++ nodeserver/namespace.js
@@ -1,6 +1,0 @@
-// namespace.js
-const namespaces = {
-    markiseStatus: 'f2etb184vq',  //
-};
-
-module.exports = namespaces;